+++ beam/animations/__init__.py
@@ -1,0 +1,8 @@
+from .langtons_ant import LangtonsAnt
+from .rain import Rain
+from .bloom import Bloom
+from .twinkle import Twinkle
+from .light import Light
+from .rainbow import Rainbow
+from .strip import Strip
+from .zap import Zap
