+++ beam/animations/bloom.py
@@ -1,0 +1,32 @@
+from animations.base import BaseBeamAnim, check_interrupt, adjustable
+import bibliopixel.colors as color_util
+from bibliopixel.util import genVector, pointOnCircle
+
+
+class Bloom(BaseBeamAnim):
+    """
+    Adapted from Maniacal labs animation lib
+    """
+
+    def __init__(self, layout, dir=True):
+        super().__init__(layout)
+        self._vector = genVector(self.layout.width, self.layout.height)
+        self._dir = dir
+
+    @check_interrupt
+    @adjustable
+    def step(self, amt=8):
+        if self._dir:
+            s = 255 - self._step
+        else:
+            s = self._step
+
+        # this respects master brightness but is slower
+        for y in range(self.layout.height):
+            for x in range(self.layout.width):
+                c = color_util.hue_helper(self._vector[y][x], self.layout.height, s)
+                self.layout.set(x, y, c)
+
+        self._step += amt
+        if self._step >= 255:
+            self._step = 0
