+++ beam/runner.py
@@ -18,7 +18,6 @@ from bibliopixel import log
 from bibliopixel.layout import Matrix
 from bibliopixel.drivers import SimPixel
 from bibliopixel.animation.animation import STATE
-from bibliopixel.util import genVector, pointOnCircle
 from bibliopixel.animation import MatrixCalibrationTest
 
 from bibliopixel.drivers.serial import Serial, LEDTYPE
@@ -41,314 +40,6 @@ def get_location(x, y):
     Sometimes useful in step() functions.
     """
     return (y * config.pixels_per_strip) + x
-
-
-class LangtonsAnt(BaseBeamAnim):
-
-    def __init__(self, layout):
-        super().__init__(layout)
-        self.offColor = color_util.Off
-        self.curColor = self.offColor
-
-    def pre_run(self):
-        self.x = random.randrange(self.width)
-        self.y = random.randrange(self.height)
-        self.d = random.randrange(4)
-
-    def __rollValue(self, val, step, _min, _max):
-        val += step
-        if val < _min:
-            diff = _min - val
-            val = _max - diff + 1
-        elif val > _max:
-            diff = val - _max
-            val = _min + diff - 1
-        return val
-
-    def __changeDir(self, direction):
-        direction = random.choice([1, -1])
-        self.d = self.__rollValue(self.d, direction, 0, 3)
-
-    def __moveAnt(self):
-        if self.d == 0:
-            self.y = self.__rollValue(self.y, 1, 0, self.height - 1)
-        elif self.d == 1:
-            self.x = self.__rollValue(self.x, 1, 0, self.width - 1)
-        elif self.d == 2:
-            self.y = self.__rollValue(self.y, -1, 0, self.height - 1)
-        elif self.d == 3:
-            self.x = self.__rollValue(self.x, -1, 0, self.width - 1)
-
-        self.curColor = self.layout.get(self.x, self.y)
-        self.layout.set(self.x, self.y, beam_state.colors[0])
-
-    @check_interrupt
-    @adjustable
-    def step(self, amt=1):
-        pathColors = beam_state.colors[1:] if len(beam_state.colors) > 1 else [color_util.Green]
-        if self.curColor in pathColors:
-            self.layout.set(self.x, self.y, self.offColor)
-            self.__changeDir(False)
-            self.__moveAnt()
-        else:
-            self.layout.set(self.x, self.y, random.choice(pathColors))
-            self.__changeDir(True)
-            self.__moveAnt()
-
-
-class ColorWipeRotate(BaseBeamAnim):
-
-    @check_interrupt
-    @adjustable
-    def step(self, amt=1):
-        for x, y in self.grid():
-            if self._step - PIXELS_PER_STRIP < x < self._step:
-                self.layout.set(x, y, random.choice(beam_state.colors))
-            else:
-                self.layout.set(x, y, color_util.Off)
-
-        if self._step >= 2 * (PIXELS_PER_STRIP - 1):
-            self._step = 0
-        else:
-            self._step += amt
-
-
-class ColorWipeSequential(BaseBeamAnim):
-
-    @check_interrupt
-    @adjustable
-    def step(self, amt=1):
-        for x, y in self.grid():
-            if get_location(x, y) < self._step:
-                self.layout.set(x, y, random.choice(beam_state.colors))
-            else:
-                self.layout.set(x, y, color_util.Off)
-
-        if self._step == PIXELS_PER_STRIP * NUM_STRIPS:
-            self._step = 0
-        else:
-            self._step += amt
-
-
-class Zap(BaseBeamAnim):
-
-    def __init__(self, layout, dir=True):
-        super().__init__(layout)
-
-        # First color
-        c_int = random.randint(0, len(beam_state.colors) - 1)
-        self.color = color = beam_state.colors[c_int]
-
-    @check_interrupt
-    @adjustable
-    def step(self, amt=1):
-        self.layout.all_off()
-
-        tail_len = 40 # Whatever
-        bullet_pos = self._step % (PIXELS_PER_STRIP + tail_len)
-
-        for x, y in self.grid():
-            if x <= bullet_pos and x > bullet_pos - tail_len:
-                beam_state.brightness = 255 - ((bullet_pos - x) * (255 / tail_len))
-                self.layout.set(x, y, color_util.color_scale(self.color, beam_state.brightness))
-
-        if bullet_pos + 1 >= PIXELS_PER_STRIP + tail_len:
-            # Sequence is about to end; Reset the Zap!
-            c_int = random.randint(0, len(beam_state.colors) - 1)
-            self.color = color = beam_state.colors[c_int]
-            self._step = 0
-        else:
-            self._step += amt
-
-
-class Twinkle(BaseBeamAnim):
-
-    def __init__(self, layout, dir=True):
-        super().__init__(layout)
-
-        self.layout = layout
-        self.colors = beam_state.colors
-        self.density = 20
-        self.speed = 2
-        self.max_bright = 255
-
-        # Make sure speed, density & max_bright are in sane ranges
-        self.speed = min(self.speed, 100)
-        self.speed = max(self.speed, 2)
-        self.density = min(self.density, 100)
-        self.density = max(self.density, 2)
-        self.max_bright = min(self.max_bright, 255)
-        self.max_bright = max(self.max_bright, 5)
-
-    def pre_run(self):
-        self._step = 0
-        # direction, color, level
-        self.pixels = [(0, color_util.Off, 0)] * self.layout.numLEDs
-
-    def pick_led(self, speed):
-        idx = random.randrange(0, self.layout.numLEDs)
-        p_dir, p_color, p_level = self.pixels[idx]
-
-        if random.randrange(0, 100) < self.density:
-            if p_dir == 0:  # 0 is off
-                p_level += speed
-                p_dir = 1  # 1 is growing
-                p_color = random.choice(self.colors)
-                self.layout._set_base(idx, color_util.color_scale(p_color, p_level))
-
-                self.pixels[idx] = p_dir, p_color, p_level
-
-    @check_interrupt
-    @adjustable
-    def step(self, amt=1):
-        self.layout.all_off()
-        self.pick_led(self.speed)
-
-        for i, val in enumerate(self.pixels):
-            p_dir, p_color, p_level = val
-            if p_dir == 1:
-                p_level += self.speed
-                if p_level > 255:
-                    p_level = 255
-                    p_dir = 2  # start dimming
-                self.layout._set_base(i, color_util.color_scale(p_color, p_level))
-            elif p_dir == 2:
-                p_level -= self.speed
-                if p_level < 0:
-                    p_level = 0
-                    p_dir = 0  # turn off
-                self.layout._set_base(i, color_util.color_scale(p_color, p_level))
-
-            self.pixels[i] = (p_dir, p_color, p_level)
-
-        self._step += amt
-
-
-class Rainbow(BaseBeamAnim):
-
-    @check_interrupt
-    @adjustable
-    def step(self, amt=1):
-        for x, y in self.grid():
-            c = color_util.wheel.wheel_color((self._step + x) % 384)
-            self.layout.set(x, y, c)
-
-        if self._step + amt == 384:
-            self._step = 0
-        else:
-            self._step += amt
-
-
-class Light(BaseBeamAnim):
-    """
-    With one color, the strip is a single-colored light. With multiple,
-    the colors just alternate along the strip (but are not animated).
-    """
-
-    @check_interrupt
-    @adjustable
-    def step(self, amt=1):
-        for x, y in self.grid():
-            c = beam_state.colors[get_location(x, y) % len(beam_state.colors)]
-            self.layout.set(x, y, c)
-
-        self._step = 0
-
-
-class Strip(BaseBeamAnim):
-    """
-    Alternate the list of colors down the strip.
-    """
-
-    @check_interrupt
-    @adjustable
-    def step(self, amt=1):
-        color_length = len(beam_state.colors)
-        for x, y in self.grid():
-            col_color = beam_state.colors[(get_location(x, y) + self._step) % color_length]
-            self.layout.set(x, y, col_color)
-
-        if self._step + amt == color_length:
-            self._step = 0
-        else:
-            self._step += amt
-
-
-class Bloom(BaseBeamAnim):
-    """
-    Adapted from Maniacal labs animation lib
-    """
-
-    def __init__(self, layout, dir=True):
-        super().__init__(layout)
-        self._vector = genVector(self.layout.width, self.layout.height)
-        self._dir = dir
-
-    @check_interrupt
-    @adjustable
-    def step(self, amt=8):
-        if self._dir:
-            s = 255 - self._step
-        else:
-            s = self._step
-
-        # this respects master brightness but is slower
-        for y in range(self.layout.height):
-            for x in range(self.layout.width):
-                c = color_util.hue_helper(self._vector[y][x], self.layout.height, s)
-                self.layout.set(x, y, c)
-
-        self._step += amt
-        if self._step >= 255:
-            self._step = 0
-
-
-class Rain(BaseBeamAnim):
-
-    def __init__(self, layout, tail=4, growth_rate=4):
-        super().__init__(layout)
-        self._tail = tail
-        self._growth_rate = growth_rate
-
-    def pre_run(self):
-        self._drops = [[] for x in range(self.layout.width)]
-
-    def _draw_drop(self, y, x, color):
-        for i in range(self._tail):
-            if x - i >= 0 and x - i < self.layout.width:
-                level = 255 - ((255 // self._tail) * i)
-                self.layout.set(x - i, y, color_util.color_scale(color, level))
-
-    @check_interrupt
-    @adjustable
-    def step(self, amt=1):
-        self.layout.all_off()
-
-        for i in range(self._growth_rate):
-            new_drop = random.randint(0, self.layout.width - 1)
-            c_int = random.randint(0, len(beam_state.colors) - 1)
-            self._drops[new_drop].append((0, beam_state.colors[c_int]))
-
-        for y in range(self.layout.height):
-            row = self._drops[y]
-            if not row:
-                pass
-
-            removals = []
-            for x in range(len(row)):
-                drop = row[x]
-                if drop[0] < self.layout.width:
-                    self._draw_drop(y, drop[0], drop[1])
-                if drop[0] - (self._tail - 1) < self.layout.width:
-                    drop = (drop[0] + 1, drop[1])
-                    self._drops[y][x] = drop
-                else:
-                    removals.append(drop)
-
-            for r in removals:
-                self._drops[y].remove(r)
-
-        self._step = 0
 
 
 Animation = enum.Enum('Animation', ' '.join(animation_dict.keys()))
@@ -458,6 +149,8 @@ if __name__ == '__main__':
         # hardware on pi
         driver = Serial(num=num_pixels, ledtype=LEDTYPE.WS2811, c_order=getattr(ChannelOrder, config.channel_order))
 
+    print(animation_dict.keys())
+    raise Exception()
     led = Matrix(
         driver,
         width=config.pixels_per_strip,
